services/video-pipeline/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
python-dotenv==1.0.0
httpx==0.25.2
